Rename utils/logging to utils/logs
