+++ backend/src/liquidity.py
@@ -3,7 +3,7 @@ import asyncio
 import aiohttp
 from aiohttp.client import ClientSession
 from tqdm.asyncio import tqdm_asyncio
-from utils.logging import *
+from utils.logs import *
 from utils.outfiles import *
 from utils.scraping import *
